NO_K_CHART mode added (if it's not defined but NO_K_LIB is defined then only KChart libraries are used but not other KDE libraries)

+++ src/analyze/gui/util.h
@@ -27,7 +27,7 @@ namespace Util {
 
 QString formatTime(qint64 ms);
 
-QString formatByteSize(int64_t size, int precision = 1);
+QString formatByteSize(double size, int precision = 1);
 
 }
